refactor: navbar는 다시 공통 컴포넌트로
