+++ src/app/components/AppSwipeNavbar.tsx
